Move TimePeriod to core package
