+++ application/src/main/java/org/opentripplanner/apis/gtfs/mapping/StreetNoteMapper.java
@@ -2,8 +2,8 @@ package org.opentripplanner.apis.gtfs.mapping;
 
 import org.opentripplanner.core.model.i18n.NonLocalizedString;
 import org.opentripplanner.core.model.id.FeedScopedId;
+import org.opentripplanner.core.model.time.TimePeriod;
 import org.opentripplanner.routing.alertpatch.Calendar;
-import org.opentripplanner.routing.alertpatch.TimePeriod;
 import org.opentripplanner.routing.alertpatch.TransitAlert;
 import org.opentripplanner.routing.alertpatch.TransitAlertBuilder;
 import org.opentripplanner.street.model.note.StreetNote;
