+++ application/src/main/java/org/opentripplanner/routing/alertpatch/Calendar.java
@@ -6,6 +6,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import org.opentripplanner.core.model.time.TimePeriod;
 import org.opentripplanner.utils.tostring.ToStringBuilder;
 
 /**
