+++ domain-core/src/test/java/org/opentripplanner/core/model/time/TimePeriodTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.alertpatch;
+package org.opentripplanner.core.model.time;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
